chore: rename 'bootstrap' command to 'import'
  Avoid clash with a few other things we call 'bootstrap' in the application already, related to the Gasket workers.
