+++ amaru/src/bin/amaru/main.rs
@@ -10,8 +10,8 @@ enum Command {
     /// Run the node in all its glory.
     Daemon(cmd::daemon::Args),
 
-    /// The ledger state from a CBOR snapshot
-    Bootstrap(cmd::bootstrap::Args),
+    /// Import the ledger state from a CBOR export produced by the Haskell node.
+    Import(cmd::import::Args),
 }
 
 #[derive(Debug, Parser)]
@@ -31,7 +31,7 @@ async fn main() -> miette::Result<()> {
 
     match args.command {
         Command::Daemon(args) => cmd::daemon::run(args).await,
-        Command::Bootstrap(args) => cmd::bootstrap::run(args).await,
+        Command::Import(args) => cmd::import::run(args).await,
     }
 }
